requirements.txt
# Environment
python-dotenv

# Google APIs (Sheets, Drive, etc.)
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# Google Gemini
google-generativeai

# Web UI
flask

# Web scraping
requests
beautifulsoup4
lxml
